feat: RandomColor 구현
- ThumbnailColor -> RandomColor로 변경
